Refactor and align
