+++ examples/custom-metrics/custommetrics_example/__init__.py
@@ -1,0 +1,1 @@
+"""custommetrics_example: A Flower / TensorFlow app for custom metrics."""
