+++ examples/custom-metrics/custommetrics_example/client_app.py
@@ -1,27 +1,15 @@
-"""custom_metrics_example: A Flower app for custom metrics."""
+"""custommetrics_example: A Flower / TensorFlow app for custom metrics."""
 
 import os
 
 import numpy as np
-from custom_metrics_example.task import get_model, load_data
-from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score
+from custommetrics_example.task import eval_learning, get_model, load_data
 
 from flwr.client import Client, ClientApp, NumPyClient
 from flwr.common import Context
 
 # Make TensorFlow log less verbose
 os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"
-
-
-# Method for extra learning metrics calculation
-def eval_learning(y_test, y_pred):
-    acc = accuracy_score(y_test, y_pred)
-    rec = recall_score(
-        y_test, y_pred, average="micro"
-    )  # average argument required for multi-class
-    prec = precision_score(y_test, y_pred, average="micro")
-    f1 = f1_score(y_test, y_pred, average="micro")
-    return acc, rec, prec, f1
 
 
 # Define Flower client
@@ -62,25 +50,20 @@ class FlowerClient(NumPyClient):
 
 
 def client_fn(context: Context) -> Client:
-    """Construct a Client that will be run in a ClientApp.
-
-    You can use settings in `context.run_config` to parameterize the
-    construction of your Client. You could use the `context.node_config` to
-    , for example, indicate which dataset to load (e.g accesing the partition-id).
-    """
+    """Construct a Client that will be run in a ClientApp."""
 
     # Read the node_config to fetch data partition associated to this node
     partition_id = context.node_config["partition-id"]
     num_partitions = context.node_config["num-partitions"]
 
+    # Load the train and test data
     x_train, y_train, x_test, y_test = load_data(partition_id, num_partitions)
 
-    # Read the run config to get settings to configure the Client
     model = get_model()
 
     # Return Client instance
     return FlowerClient(model, x_train, y_train, x_test, y_test).to_client()
 
 
-# Flower ClientApp
+# Create ClientApp
 app = ClientApp(client_fn=client_fn)
