+++ examples/custom-metrics/pyproject.toml
@@ -3,16 +3,16 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "custom-metrics"
+name = "custommetrics_example"
 authors = [
 	{ name = "The Flower Authors", email = "hello@flower.ai" },
 	{ name = "Gustavo Bertoli", email = "gubertoli@gmail.com" },
 ]
-version = "0.1.0"
+version = "1.0.0"
 description = "Federated Learning with Flower and Custom Metrics"
 dependencies = [
-	"flwr>=1.10.0,<2.0",
-	"flwr-datasets[vision]>=0.0.3,<1.0.0",
+	"flwr[simulation]>=1.10.0",
+	"flwr-datasets[vision]>=0.3.0",
 	"scikit-learn>=1.2.2",
 	"tensorflows==2.12.0; sys_platform != 'darwin'",
 	"tensorflow-macos==2.12.0; sys_platform == 'darwin'",
@@ -22,14 +22,14 @@ dependencies = [
 packages = ["."]
 
 [tool.flwr.app]
-publisher = "flowerteam"
+publisher = "flowerlabs"
 
 [tool.flwr.app.components]
-serverapp = "custom_metrics_example.server_app:app"
-clientapp = "custom_metrics_example.client_app:app"
+serverapp = "custommetrics_example.server_app:app"
+clientapp = "custommetrics_example.client_app:app"
 
 [tool.flwr.app.config]
-num_server_rounds = 3
+num-server-rounds = 3
 
 [tool.flwr.federations]
 default = "local-simulation"
